refactor(onboarding): step-based state machine, animation hook, auth flow
Step-based refactor:
- Replace 6 boolean states with single OnboardingStep enum
- Simplify body overflow, engagement animation, JSX conditionals
- Split routing into query param handler + logged-in redirect
- Remove dead code: importFlowSourceRef, onbSignup effect, 'manual' context

Animation extraction:
- Create useOnboardingAnimations hook (mount, tags, feed, parallax,
  hero tracking, engagement floater, confetti, browser detection, body overflow)

Auth flow fixes:
- Add autoTriggerProvider prop to AuthOptionsProps for GitHub auto-trigger
- onSuccessfulRegistration triggers import based on signupContext
- Routing effect guarded by step (blocks during auth/importing/extension/complete)
- Completion actions fire on seniority selection, not prematurely
- Experience level saved via UPDATE_USER_PROFILE_MUTATION on seniority pick
